moved description box into its own fragment
