+++ app/src/main/res/layout/fragment_my_posts.xml
@@ -4,11 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="post"
-        tools:layout_editor_absoluteX="71dp"
-        tools:layout_editor_absoluteY="184dp" />
+    <include layout="@layout/fragment_temp_feed"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
